End-of-Day Commit
I successfully updated method $configure() as well updated method $set.q_graphs().  The latter has a different structure for the 'data' vertex attribute; therefore, functions make.evs_universe() and evs_retrace() need to be updated to accomodate this.
For make.evs_universe(), it appears that the only change needed is to update the definition of the internal object '.tmp_space' to retrieve source data using quosures.